Refactor merge viewer codemirror setup
